(김장성) 보호소 파일 업로드 기능
FileDownloadView.java - 파일 다운로드를 위한 클래스 생성

BusinessLicenseCommand.java - view에서 데이터와 파일을 받기위한 클래스 생성

BusinessLicenseFile.java - t_overall_file DTO 생성

ShelterMapper.xml
 selectOfCodeNum - t_overall_file 테이블의 of_code를 잘라서 가장 높은 숫자를 리턴하는 쿼리문
추가

ShelterDao.java
 businessLicense() - MVC 규칙을 위한 businessLicense 메서드
 selectOfCodeNum() - t_overall_file 테이블의 of_code를 잘라서 가장 높은 숫자를 리턴하는
쿼리문을 호출하는 메서드

ShelterService.java
 addBusinessLicense(...) - 보호소 대표 신청 메서드에 파일 등록 기능 추가
 businessLicense() - MVC 규칙을 위한 businessLicense 메서드

ShelterController.java
 BusinessLicense() - MVC 규칙 유지하기위해 메서드 호출 추가
 addBusinessLicense(...) - MultipartFile 매개변수 추가 및 BusinessLicense ->
BusinessLicenseCommand로 매개변수 변경, 리소스 경로 입력받는 변수 추가

html5BasicTemplate.jsp - 템플릿에 부트스트랩,제이쿼리 추가

businessLicenseRequest.jsp - 파일 업로드 추가(multipartFile)

shelterMenu.jsp - 보호소 기능 메뉴 추가

+++ teamProject/src/main/java/com/cafe24/iso159/shelter/service/BusinessLicenseFile.java
@@ -7,7 +7,7 @@ public class BusinessLicenseFile {
 	private String ofOriginName;
 	private String ofSaveName;
 	private String ofExt;
-	private int ofSize;
+	private long ofSize;
 	private String ofDate;
 	public String getOfCode() {
 		return ofCode;
@@ -45,10 +45,10 @@ public class BusinessLicenseFile {
 	public void setOfExt(String ofExt) {
 		this.ofExt = ofExt;
 	}
-	public int getOfSize() {
+	public long getOfSize() {
 		return ofSize;
 	}
-	public void setOfSize(int ofSize) {
+	public void setOfSize(long ofSize) {
 		this.ofSize = ofSize;
 	}
 	public String getOfDate() {
@@ -62,6 +62,7 @@ public class BusinessLicenseFile {
 		return "BusinessLicenseFile [ofCode=" + ofCode + ", blCode=" + blCode + ", ofPath=" + ofPath + ", ofOriginName="
 				+ ofOriginName + ", ofSaveName=" + ofSaveName + ", ofExt=" + ofExt + ", ofSize=" + ofSize + ", ofDate="
 				+ ofDate + "]";
-	}	
+	}
+
 	
 }
